Module traffloat::cargo

Management of cargo in buildings

Structs

NextStorageSize

The size of a cargo storage in the next simulation frame.

Storage

A component attached to storage entities.

StorageCapacity

A component attached to nodes to inidcate cargo capacity of the node.

StorageList

A component attached to nodes to indicate cargo in the node.

StorageSize

The size of a cargo storage in the current simulation frame.

Functions

lerp

Interpolates the current graphical size of a storage.

setup_ecs

Initializes ECS

update_storage

Copy next.size into current.size

update_storage_setup

update_storage_system