Module traffloat::edge

Edge management.

An edge is also called a “corridor”. It connects two nodes together.

Structs

AddEvent

Indicates that an edge is added.

CreateRequest

An event to schedule requests to initialize new edges.

CrossSectionPosition

A position on the cross section of an edge.

Design

The geometric design of the edge.

Duct

A circular structure in an edge.

ElectricityDuctType

Details of a electricity duct.

Id

Component storing the endpoints of an edge

LoadRequest

An event to schedule requests to initialize saved edges.

RailDuctType

Details of a rail duct.

RemoveEvent

Indicates that an edge is flagged for removal

Size

Defines the size of an edge

Enums

Direction

A direction across an edge.

DuctType

The type of a duct.

LiquidDuctType

Details of a liquid duct.

Functions

create_duct

Creates a duct entity for a duct type.

create_new_edge

create_new_edge_setup

create_new_edge_system

create_saved_edge

create_saved_edge_setup

create_saved_edge_system

save_edges

save_edges_setup

save_edges_system

setup_ecs

Initializes ECS

tf

Computes the transformation matrix from or to the unit cylinder