Module traffloat::factory::resource

Wrapper for factory resources

Structs

Cargo

This reaction requires cargo.

Gas

This reaction requires gas.

Light

This reaction requires sunlight.

Liquid

This reaction requires liquid.

Power

This reaction requires power.

Skill

This reaction requires an operator with the specified skill.

Enums

Conditional

A type of resource whose existence affects a reaction.

Consumable

A type of resource that can be consumed.