Module traffloat::gas

Management of gas in buildings

Structs

NextStorageSize

The size of a gas storage in the next simulation frame.

Pump

A component applied on a node that drives gas.

Storage

A component attached to storage entities.

StorageCapacity

A component attached to nodes to inidcate cargo capacity of the node.

StorageList

A component attached to entities that house gas.

StorageSize

The size of a gas storage in the current simulation frame.

Functions

lerp

Interpolates the current graphical size of a storage.

setup_ecs

Initializes ECS

update_storage

update_storage_setup

update_storage_system