Module traffloat::node

Node management.

A node is an instance of a building.

Structs

AddEvent

Indicates that a node is added

Child

A component applied to child entities of a node.

CreationRequest

An event to schedule requests to initialize new nodes.

Id

Component storing a persistent identifier for a node

Index

Tracks the nodes in the world

LoadRequest

An event to schedule requests to initialize saved nodes.

PostRemoveEvent

Indicates that nodes have been removed

RemoveEvent

Indicates that a node is flagged for removal

Functions

create_new_node

create_new_node_setup

create_new_node_system

create_saved_node

create_saved_node_setup

create_saved_node_system

delete_nodes

delete_nodes_setup

delete_nodes_system

init_features

save_nodes

save_nodes_setup

save_nodes_system

setup_ecs

Initializes ECS