Module traffloat::space

Spatial units

Structs

Position

A component storing the world position of an object.

Functions

transform_cuboid

Creates a transformation matrix from a cube to a cuboid at lower..upper.

transform_cylinder

Creates a transformation matrix from a unit cylinder cuboid to one with base center (0, 0, -zn) and top center (0, 0, zn), with cross-section ellipse radii x, y on the X, Y axes.

Type Definitions

LinearMatrix

Standard linear transformation matrix type

Matrix

Standard homogenous matrix type

Point

Standard vector type

Vector

Standard vector type