Module traffloat::node

Node management.

A node is an instance of a building.

Modules

save

Save type for nodes.

Structs

AddEvent

Indicates that a node is added

Child

A component applied to child entities of a node.

CreationRequest

An event to schedule requests to initialize new nodes.

Id

Component storing an identifier for a node

Index

Tracks the nodes in the world

LoadRequest

An event to schedule requests to initialize saved nodes.

Name

Component storing the name of the node

PostRemoveEvent

Indicates that nodes have been removed

RemoveEvent

Indicates that a node is flagged for removal

Functions

create_new_node

create_new_node_setup

create_new_node_system

create_saved_node

create_saved_node_setup

create_saved_node_system

delete_nodes

delete_nodes_setup

delete_nodes_system

setup_ecs

Initializes ECS